finish final tbl

+++ public/js/custom.js
@@ -184,10 +184,16 @@ $(document).ready(function(){
 	if (total > 0) {
 		$('#total').html(`Total: $${total}`).css({"font-size": "25px", "color": "red"});
 		$('#cartTotal').val(total);
+		$('#cartdetail').val(JSON.stringify(dataList));
 	}
 	$('#cartForm').submit(function(){
-		alert("Your order has been sent!");
-		sessionStorage.clear();
+		if (confirm('Check your information again, if you are sure then press OK')) {
+			alert("Your order has been sent!");
+			sessionStorage.clear();
+			return true;
+		} else {
+			return false;
+		}
 	});
 	total = 0;
